git-log: add function to view file from revision, prompt if necessary.
Since the inherited log-view-mode is so broken for git. Attempted to make it
slightly less broken with vars telling it what we're doing, but I think I'll
just axe its menu.

Small improvement in error robustness for git--cat-file.

Small tweaks for the command-line-triggered git log to work well on Mac.