More auto smarts for git-merge-next-action and maybe-ask-commit.
git-merge-next-action and a few other commands that may apply to a remote
automatically ask for a remote refresh if the tracking branch seems old;
this is configurable but defaults to a day. A new utility function
git--maybe-ask-for-fetch is used around branch names that could be remotes.

maybe-ask commit now also offers stash as an option.

Fix corner case of switching branch off of detached HEAD.