Add commands for git push and pull using the new async run mode.
Delete the previous -ff-only, which were inflexible and a little too
complex. The new commands ask the user for a complete command, using
history rather than completion for quickness -- it's my experience that I
mostly push and pull the exact same things; moreover, git is the better
place to configure defaults than emacs.

They accomodate things like dry-run easily, and will not block input in
case of expensive hooks like compilation/tests. Because they run in a
comint buffer, they are able to handle password prompts gracefully. We also
have more space to display information, and in the case of pull we use it
to give the user a quick link back to the previous HEAD.

The goal is for most of the flow to occur in emacs not terminal, even when
I expect complex commands.