After doing a textual diff buffer, set default-directory to repo root.
At some undetermined point in the past, I think git diff switched to
producing paths relative to repo root. If the default-directory is
different than that actual base name, diff-mode will not be able to
find the sources for hunk and visit operations.

This does not affect where we run `git diff`, just the subsequent
operation of diff-mode.